[python-frontend] reduce preprocessor facade to composed mixins and minimal initializer
